infrastructure/containers/overlord/runtime/requirements.txt
# Lucid overlord services — minimal API for env/config orchestration (Python 3.11)
fastapi>=0.109.0,<0.115.0
uvicorn[standard]>=0.27.0,<0.31.0
PyYAML>=6.0.1,<7
